Allow more than two shapes to be drawn on the screen
Changed the shape array to an ArrayList. Added a printShapeArray status
to show what kinds of shapes are in the ArrayList - this uses string
manipulation of the toString methods of the Shape objects. Modified
compare to methods to accomodate taking any two shapes, but commented
out the code that calls the comparison.